+++ src/Asymmetric/SecretKey.php
@@ -17,7 +17,7 @@ class SecretKey extends Key
     public function __construct(HiddenString $keyMaterial)
     {
         parent::__construct($keyMaterial);
-        $this->is_asymmetric_key = true;
+        $this->isAsymmetricKey = true;
     }
     
     /**