+++ src/SignatureKeyPair.php
@@ -20,15 +20,14 @@ final class SignatureKeyPair extends KeyPair
     /**
      * @var SignatureSecretKey
      */
-    protected $secret_key;
+    protected $secretKey;
 
     /**
      * @var SignaturePublicKey
      */
-    protected $public_key;
+    protected $publicKey;
 
     /**
-     * 
      * Pass it a secret key, it will automatically generate a public key
      * 
      * @param Key[] $keys
@@ -112,8 +111,8 @@ final class SignatureKeyPair extends KeyPair
      */
     protected function setupKeyPair(SignatureSecretKey $secret)
     {
-        $this->secret_key = $secret;
-        $this->public_key = $this->secret_key->derivePublicKey();
+        $this->secretKey = $secret;
+        $this->publicKey = $this->secretKey->derivePublicKey();
     }
 
     /**
@@ -123,8 +122,9 @@ final class SignatureKeyPair extends KeyPair
      */
     public function getPublicKey()
     {
-        return $this->public_key;
+        return $this->publicKey;
     }
+
     /**
      * Get a Key object for the public key
      *
@@ -132,6 +132,6 @@ final class SignatureKeyPair extends KeyPair
      */
     public function getSecretKey()
     {
-        return $this->secret_key;
+        return $this->secretKey;
     }
 }